Fix `codex_path` handling
If the user provided a string handle it with `shlex.split`.  In case
the first part of it, refer to it as executable, is not given as an
absolute path, use `shutil.which` to resolve it.

This allows (1) short names like "codex" to be expanded to its full
path, but also (2) complex commands like "sh --login -i -c codex" to
get split and resolved automatically.  E.g. the latter could get trans-
formed to `["C:\\Dev\\Git\\bin\\sh.exe", "--login", "-i", "codex"]`.